# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Authentication and authorization REST API (JWT access + refresh, OTP, RBAC, registration invites).",
        "title": "Auth service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/admin/signing-keys/rotate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Rotate signing key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Required for human access tokens; omitted for verified service tokens",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Case-insensitive login or email substring",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque keyset cursor; omit on first request",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.AdminUsersResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/admin/users/{userID}/ban": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Ban user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Required for human access tokens; omitted for verified service tokens",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    },
                    {
                        "description": "Reason",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.BanUserRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Unban user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User UUID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Required for human access tokens; omitted for verified service tokens",
                        "name": "X-CSRF-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/auth/has-role": {
            "post": {
                "description": "Intended for another service authorizing a human caller. The user ID is derived only from the verified access token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Check a role using an access token from JSON",
                "parameters": [
                    {
                        "description": "Access token and role name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.TokenHasRoleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.HasRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
        "/v1/auth/has-role-with-tag": {
            "post": {
                "description": "Intended for another service authorizing a human caller. Role and tag inheritance use the normal authorization rules.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Check a role tag using an access token from JSON",
                "parameters": [
                    {
                        "description": "Access token, role name, and tag",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.TokenHasRoleWithTagRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/httptransport.HasRoleWithTagResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
...
        "/v1/auth/service-token": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Issue service access token",
                "parameters": [
                    {
                        "description": "Service login and secret",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httptransport.ServiceTokenRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "access_token only (no csrf_token)",
                        "schema": {
                            "$ref": "#/definitions/httptransport.TokenPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httptransport.ErrEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httptransport.AdminUsersResponse": {
            "type": "object",
            "properties": {
                "next_cursor": {
                    "type": "string"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer",
                    "x-nullable": true
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httptransport.AdminUserRow"
                    }
                }
            }
        },
...
        "httptransport.BanUserRequest": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "httptransport.ErrEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "httptransport.HasRoleResponse": {
            "type": "object",
            "properties": {
                "has_role": {
                    "type": "boolean"
                }
            }
        },
        "httptransport.HasRoleWithTagResponse": {
            "type": "object",
            "properties": {
                "has_role_with_tag": {
                    "type": "boolean"
                }
            }
        },
...
        "httptransport.TokenHasRoleRequest": {
            "type": "object",
            "properties": {
                "role_name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "httptransport.TokenHasRoleWithTagRequest": {
            "type": "object",
            "properties": {
                "role_name": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the access JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}